+++ examples/018-composer.rb
@@ -3,9 +3,10 @@
 # This example shows how [HexaPDF::Composer] simplifies the creation of PDF
 # documents by providing a high-level interface to the box layouting engine.
 #
-# Basic style properties can be set on the [HexaPDF::Composer#base_style] style.
-# These properties are reused by every box and can be adjusted on a box-by-box
-# basis.
+# Basic style properties can be set using the [HexaPDF::Composer#style] method
+# and the style name `:basic`. These properties are reused by every box and can
+# be adjusted on a box-by-box basis. Newly defined styles also inherit the
+# properties from the `:basic` style.
 #
 # Various methods allow the easy creation of boxes, for example, text and image
 # boxes. All these boxes are automatically drawn on the page. If the page has
@@ -24,21 +25,20 @@ dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exer\u{00AD}citation
 ullamco laboris nisi ut aliquip ex ea commodo consequat. ".tr("\n", " ")
 
 HexaPDF::Composer.create('composer.pdf') do |pdf|
-  pdf.base_style.update(line_spacing: {type: :proportional, value: 1.5},
-                        last_line_gap: true, align: :justify)
-  image_style = pdf.base_style.dup.update(border: {width: 1}, padding: 5, margin: 10)
-  link_style = pdf.base_style.dup.update(fill_color: [6, 158, 224], underline: true)
+  pdf.style(:base, line_spacing: 1.5, last_line_gap: true, align: :justify)
+  pdf.style(:image, border: {width: 1}, padding: 5, margin: 10)
+  pdf.style(:link, fill_color: [6, 158, 224], underline: true)
   image = File.join(__dir__, 'machupicchu.jpg')
 
   pdf.text(lorem_ipsum * 2)
-  pdf.image(image, style: image_style, width: 200, position: :float)
-  pdf.image(image, style: image_style, width: 200, position: :absolute,
+  pdf.image(image, style: :image, width: 200, position: :float)
+  pdf.image(image, style: :image, width: 200, position: :absolute,
             position_hint: [200, 300])
   pdf.text(lorem_ipsum * 20, position: :flow)
 
   pdf.formatted_text(["Produced by ",
                       {link: "https://hexapdf.gettalong.org", text: "HexaPDF",
-                       style: link_style},
+                       style: :link},
                       " via HexaPDF::Composer"],
                       font_size: 15, align: :center, padding: 15)
 end